refactor: move metrics module to playbook module
